+++ server/src-lib/Hasura/GraphQL/Execute/Mutation.hs
@@ -105,9 +105,6 @@ convertMutationAction
     , MonadError QErr m
     , MonadMetadataStorage (MetadataStorageT m)
     , Tracing.MonadTrace m
-    , Tracing.MonadTrace tx
-    , MonadIO tx
-    , MonadTx tx
     )
   => Env.Environment
   -> L.Logger L.Hasura
@@ -115,12 +112,13 @@ convertMutationAction
   -> HTTP.Manager
   -> HTTP.RequestHeaders
   -> ActionMutation 'Postgres (UnpreparedValue 'Postgres)
-  -> m (tx EncJSON, HTTP.ResponseHeaders)
+  -> m (ActionExecutionPlan, HTTP.ResponseHeaders)
 convertMutationAction env logger userInfo manager reqHeaders = \case
-  AMSync s  -> ((unActionExecuteTx . _aerExecution) &&& _aerHeaders) <$>
+  AMSync s  -> ((AEPSync . _aerExecution) &&& _aerHeaders) <$>
     resolveActionExecution env logger userInfo s actionExecContext
-  AMAsync s -> (noResponseHeaders . unActionExecuteTx) <$>
-               liftEitherM (runMetadataStorageT $ resolveActionMutationAsync s reqHeaders userSession)
+  AMAsync s -> do
+    result <- liftEitherM (runMetadataStorageT $ resolveActionMutationAsync s reqHeaders userSession)
+    pure (AEPAsyncMutation result, [])
   where
     userSession = _uiSession userInfo
     actionExecContext = ActionExecContext manager reqHeaders $ _uiSession userInfo
@@ -146,7 +144,7 @@ convertMutationSelectionSet
   -> G.SelectionSet G.NoFragments G.Name
   -> [G.VariableDefinition]
   -> Maybe GH.VariableValues
-  -> m (ExecutionPlan (tx EncJSON, HTTP.ResponseHeaders))
+  -> m (ExecutionPlan (ActionExecutionPlan, HTTP.ResponseHeaders) (tx EncJSON, HTTP.ResponseHeaders))
 convertMutationSelectionSet env logger gqlContext SQLGenCtx{stringifyNum} userInfo manager reqHeaders fields varDefs varValsM = do
   mutationParser <- onNothing (gqlMutationParser gqlContext) $
     throw400 ValidationFailed "no mutations exist"
@@ -160,7 +158,7 @@ convertMutationSelectionSet env logger gqlContext SQLGenCtx{stringifyNum} userIn
   let userSession = _uiSession userInfo
       remoteJoinCtx = (manager, reqHeaders, userInfo)
   txs <- for unpreparedQueries \case
-    RFDB db -> ExecStepDB . noResponseHeaders <$> case db of
+    RFDB _ execCtx db -> ExecStepDB execCtx . noResponseHeaders <$> case db of
       MDBInsert s   -> convertInsert env userSession remoteJoinCtx s stringifyNum
       MDBUpdate s   -> convertUpdate env userSession remoteJoinCtx s stringifyNum
       MDBDelete s   -> convertDelete env userSession remoteJoinCtx s stringifyNum
@@ -172,7 +170,7 @@ convertMutationSelectionSet env logger gqlContext SQLGenCtx{stringifyNum} userIn
         remoteSchemaInfo
         G.OperationTypeMutation
         $ [G.SelectionField resolvedRemoteField]
-    RFAction action     -> ExecStepDB <$> convertMutationAction env logger userInfo manager reqHeaders action
+    RFAction action     -> ExecStepAction <$> convertMutationAction env logger userInfo manager reqHeaders action
     RFRaw s             -> pure $ ExecStepRaw s
 
   return txs