+++ server/src-lib/Hasura/RQL/DML/Count.hs
@@ -11,10 +11,12 @@ import           Hasura.Prelude
 import qualified Data.ByteString.Builder                    as BB
 import qualified Data.Sequence                              as DS
 
+import           Control.Monad.Trans.Control                (MonadBaseControl)
 import           Data.Aeson
 
 import qualified Database.PG.Query                          as Q
 import qualified Hasura.Backends.Postgres.SQL.DML           as S
+import qualified Hasura.Tracing                             as Tracing
 
 import           Hasura.Backends.Postgres.SQL.Types
 import           Hasura.Backends.Postgres.Translate.BoolExp
@@ -23,6 +25,7 @@ import           Hasura.RQL.DML.Internal
 import           Hasura.RQL.DML.Types
 import           Hasura.RQL.IR.BoolExp
 import           Hasura.RQL.Types
+import           Hasura.RQL.Types.Run
 import           Hasura.SQL.Types
 
 
@@ -65,13 +68,13 @@ mkSQLCount (CountQueryP1 tn (permFltr, mWc) mDistCols) =
 -- SELECT count(*) FROM (SELECT DISTINCT c1, .. cn FROM .. WHERE ..) r;
 -- SELECT count(*) FROM (SELECT * FROM .. WHERE ..) r;
 validateCountQWith
-  :: (UserInfoM m, QErrM m, CacheRM m)
+  :: (UserInfoM m, QErrM m, TableInfoRM 'Postgres m)
   => SessVarBldr 'Postgres m
   -> (ColumnType 'Postgres -> Value -> m S.SQLExp)
   -> CountQuery
   -> m CountQueryP1
 validateCountQWith sessVarBldr prepValBldr (CountQuery qt mDistCols mWhere) = do
-  tableInfo <- askTabInfo qt
+  tableInfo <- askTabInfoSource qt
 
   -- Check if select is allowed
   selPerm <- modifyErr (<> selNecessaryMsg) $
@@ -105,9 +108,11 @@ validateCountQWith sessVarBldr prepValBldr (CountQuery qt mDistCols mWhere) = do
 
 validateCountQ
   :: (QErrM m, UserInfoM m, CacheRM m)
-  => CountQuery -> m (CountQueryP1, DS.Seq Q.PrepArg)
-validateCountQ =
-  runDMLP1T . validateCountQWith sessVarFromCurrentSetting binRHSBuilder
+  => SourceName -> CountQuery -> m (CountQueryP1, DS.Seq Q.PrepArg)
+validateCountQ source query = do
+  tableCache <- askTableCache source
+  flip runTableCacheRT (source, tableCache) $ runDMLP1T $
+    validateCountQWith sessVarFromCurrentSetting binRHSBuilder query
 
 countQToTx
   :: (QErrM m, MonadTx m)
@@ -122,7 +127,11 @@ countQToTx (u, p) = do
       BB.byteString "{\"count\":" <> BB.intDec c <> BB.char7 '}'
 
 runCount
-  :: (QErrM m, UserInfoM m, CacheRM m, MonadTx m)
-  => CountQuery -> m EncJSON
-runCount q =
-  validateCountQ q >>= countQToTx
+  :: ( QErrM m, UserInfoM m, CacheRM m
+     , MonadIO m, MonadBaseControl IO m
+     , Tracing.MonadTrace m
+     )
+  => SourceName -> CountQuery -> m EncJSON
+runCount source q = do
+  sourceConfig <- _pcConfiguration <$> askPGSourceCache source
+  validateCountQ source q >>=  liftEitherM . runExceptT . runQueryLazyTx (_pscExecCtx sourceConfig) Q.ReadOnly . countQToTx