+++ server/src-lib/Hasura/RQL/Types/Common.hs
@@ -56,42 +56,46 @@ module Hasura.RQL.Types.Common
        , UrlConf(..)
        , resolveUrlConf
        , getEnv
+
+       , SourceName(..)
+       , defaultSource
+       , sourceNameToText
        ) where
 
 import           Hasura.Prelude
 
-import qualified Data.Environment                   as Env
-import qualified Data.HashMap.Strict                as HM
-import qualified Data.Text                          as T
-import qualified Database.PG.Query                  as Q
-import qualified Language.GraphQL.Draft.Syntax      as G
-import qualified Language.Haskell.TH.Syntax         as TH
-import qualified PostgreSQL.Binary.Decoding         as PD
-import qualified Test.QuickCheck                    as QC
+import qualified Data.Environment                       as Env
+import qualified Data.HashMap.Strict                    as HM
+import qualified Data.Text                              as T
+import qualified Database.PG.Query                      as Q
+import qualified Language.GraphQL.Draft.Syntax          as G
+import qualified Language.Haskell.TH.Syntax             as TH
+import qualified PostgreSQL.Binary.Decoding             as PD
+import qualified Test.QuickCheck                        as QC
 
-import           Control.Lens                       (makeLenses)
+import           Control.Lens                           (makeLenses)
 import           Data.Aeson
 import           Data.Aeson.Casing
 import           Data.Aeson.TH
-import           Data.Bifunctor                     (bimap)
-import           Data.Kind                          (Type)
-import           Data.Scientific                    (toBoundedInteger)
+import           Data.Bifunctor                         (bimap)
+import           Data.Kind                              (Type)
+import           Data.Scientific                        (toBoundedInteger)
 import           Data.Text.Extended
 import           Data.Text.NonEmpty
 import           Data.Typeable
 import           Data.URL.Template
 
-import qualified Hasura.Backends.Postgres.SQL.DML   as PG
-import qualified Hasura.Backends.Postgres.SQL.Types as PG
-import qualified Hasura.Backends.Postgres.SQL.Value as PG
+import qualified Hasura.Backends.Postgres.Execute.Types as PG
+import qualified Hasura.Backends.Postgres.SQL.DML       as PG
+import qualified Hasura.Backends.Postgres.SQL.Types     as PG
+import qualified Hasura.Backends.Postgres.SQL.Value     as PG
 
 import           Hasura.EncJSON
-import           Hasura.Incremental                 (Cacheable)
-import           Hasura.RQL.DDL.Headers             ()
+import           Hasura.Incremental                     (Cacheable)
+import           Hasura.RQL.DDL.Headers                 ()
 import           Hasura.RQL.Types.Error
 import           Hasura.SQL.Backend
 import           Hasura.SQL.Types
-
 
 type Representable a = (Show a, Eq a, Hashable a, Cacheable a, NFData a)
 
@@ -162,6 +166,7 @@ class
   type XAILIKE        b :: Type
   type XANILIKE       b :: Type
   type XComputedFieldInfo b :: Type
+  type SourceConfig   b :: Type
   isComparableType :: ScalarType b -> Bool
   isNumType :: ScalarType b -> Bool
 
@@ -182,6 +187,7 @@ instance Backend 'Postgres where
   type XAILIKE        'Postgres = ()
   type XANILIKE       'Postgres = ()
   type XComputedFieldInfo 'Postgres = ()
+  type SourceConfig   'Postgres = PG.PGSourceConfig
   isComparableType = PG.isComparableType
   isNumType = PG.isNumType
 
@@ -286,21 +292,57 @@ fromRel = FieldName . relNameToTxt
 class ToAesonPairs a where
   toAesonPairs :: (KeyValue v) => a -> [v]
 
+data SourceName
+  = SNDefault
+  | SNName !NonEmptyText
+  deriving (Show, Eq, Ord, Generic)
+
+instance FromJSON SourceName where
+  parseJSON = withText "String" $ \case
+    "default" -> pure SNDefault
+    t         -> SNName <$> parseJSON (String t)
+
+sourceNameToText :: SourceName -> Text
+sourceNameToText = \case
+  SNDefault -> "default"
+  SNName t  -> unNonEmptyText t
+
+instance ToJSON SourceName where
+  toJSON = String . sourceNameToText
+
+instance ToTxt SourceName where
+  toTxt = sourceNameToText
+
+instance ToJSONKey SourceName
+instance Hashable SourceName
+instance NFData SourceName
+instance Cacheable SourceName
+
+instance Arbitrary SourceName where
+  arbitrary = SNName <$> arbitrary
+
+defaultSource :: SourceName
+defaultSource = SNDefault
+
 data WithTable a
   = WithTable
-  { wtName :: !PG.QualifiedTable
-  , wtInfo :: !a
+  { wtSource :: !SourceName
+  , wtName   :: !PG.QualifiedTable
+  , wtInfo   :: !a
   } deriving (Show, Eq)
 
 instance (FromJSON a) => FromJSON (WithTable a) where
   parseJSON v@(Object o) =
-    WithTable <$> o .: "table" <*> parseJSON v
+    WithTable
+    <$> o .:? "source" .!= defaultSource
+    <*> o .: "table"
+    <*> parseJSON v
   parseJSON _ =
     fail "expecting an Object with key 'table'"
 
 instance (ToAesonPairs a) => ToJSON (WithTable a) where
-  toJSON (WithTable tn rel) =
-    object $ ("table" .= tn):toAesonPairs rel
+  toJSON (WithTable sourceName tn rel) =
+    object $ ("source" .= sourceName):("table" .= tn):toAesonPairs rel
 
 type ColumnValues a = HM.HashMap PG.PGCol a