+++ server/src-lib/Hasura/Server/API/Query.hs
@@ -120,7 +120,7 @@ data RQLQueryV1
 
   | RQRunSql !RunSQL
 
-  | RQReplaceMetadata !Metadata
+  | RQReplaceMetadata !ReplaceMetadata
   | RQExportMetadata !ExportMetadata
   | RQClearMetadata !ClearMetadata
   | RQReloadMetadata !ReloadMetadata
@@ -184,19 +184,26 @@ $(deriveJSON
 runQuery
   :: ( HasVersion, MonadIO m, Tracing.MonadTrace m
      , MonadBaseControl IO m, MonadMetadataStorage m
+     , MonadResolveSource m
      )
-  => Env.Environment -> PGExecCtx -> InstanceId
+  => Env.Environment
+  -> InstanceId
   -> UserInfo -> RebuildableSchemaCache -> HTTP.Manager
   -> SQLGenCtx -> RemoteSchemaPermsCtx -> RQLQuery -> m (EncJSON, RebuildableSchemaCache)
-runQuery env pgExecCtx instanceId userInfo sc hMgr sqlGenCtx remoteSchemaPermsCtx query = do
-  accessMode <- getQueryAccessMode query
-  traceCtx <- Tracing.currentContext
+runQuery env instanceId userInfo sc hMgr sqlGenCtx remoteSchemaPermsCtx query = do
   metadata <- fetchMetadata
-  result <- runQueryM env query & Tracing.interpTraceT \x -> do
+  let sources = scPostgres $ lastBuiltSchemaCache sc
+
+  (sourceName, _) <- case HM.toList sources of
+    []  -> throw400 NotSupported "no postgres source exist"
+    [s] -> pure $ second _pcConfiguration s
+    _   -> throw400 NotSupported "multiple postgres sources found"
+
+  result <- runQueryM env sourceName query & Tracing.interpTraceT \x -> do
     (((js, tracemeta), meta), rsc, ci) <-
          x & runMetadataT metadata
            & runCacheRWT sc
-           & peelRun runCtx pgExecCtx accessMode (Just traceCtx)
+           & peelRun runCtx
            & runExceptT
            & liftEitherM
     pure ((js, rsc, ci, meta), tracemeta)
@@ -346,17 +353,19 @@ reconcileAccessModes (Just mode1) (Just mode2)
   | otherwise = Left mode2
 
 runQueryM
-  :: ( HasVersion, QErrM m, CacheRWM m, UserInfoM m, MonadTx m
-     , MonadIO m, MonadUnique m, HasHttpManager m, HasSQLGenCtx m
+  :: ( HasVersion, CacheRWM m, UserInfoM m
+     , MonadBaseControl IO m, MonadIO m, MonadUnique m
+     , HasHttpManager m, HasSQLGenCtx m
      , HasRemoteSchemaPermsCtx m
      , Tracing.MonadTrace m
      , MetadataM m
-     , MonadScheduledEvents m
+     , MonadMetadataStorageQueryAPI m
      )
   => Env.Environment
+  -> SourceName
   -> RQLQuery
   -> m EncJSON
-runQueryM env rq = withPathK "args" $ case rq of
+runQueryM env source rq = withPathK "args" $ case rq of
   RQV1 q -> runQueryV1M q
   RQV2 q -> runQueryV2M q
   where
@@ -392,11 +401,11 @@ runQueryM env rq = withPathK "args" $ case rq of
       RQGetInconsistentMetadata q     -> runGetInconsistentMetadata q
       RQDropInconsistentMetadata q    -> runDropInconsistentMetadata q
 
-      RQInsert q                      -> runInsert env q
-      RQSelect q                      -> runSelect q
-      RQUpdate q                      -> runUpdate env q
-      RQDelete q                      -> runDelete env q
-      RQCount  q                      -> runCount q
+      RQInsert q                      -> runInsert env source q
+      RQSelect q                      -> runSelect source q
+      RQUpdate q                      -> runUpdate env source q
+      RQDelete q                      -> runDelete env source q
+      RQCount  q                      -> runCount source q
 
       RQAddRemoteSchema    q          -> runAddRemoteSchema env q
       RQRemoveRemoteSchema q          -> runRemoveRemoteSchema q
@@ -440,18 +449,17 @@ runQueryM env rq = withPathK "args" $ case rq of
 
       RQDumpInternalState q           -> runDumpInternalState q
 
-      RQRunSql q                      -> runRunSQL q
+      RQRunSql q                      -> runRunSQL defaultSource q
 
       RQSetCustomTypes q              -> runSetCustomTypes q
       RQSetTableCustomization q       -> runSetTableCustomization q
 
-      RQBulk qs                       -> encJFromList <$> indexedMapM (runQueryM env) qs
+      RQBulk qs                       -> encJFromList <$> indexedMapM (runQueryM env source) qs
 
     runQueryV2M = \case
       RQV2TrackTable q           -> runTrackTableV2Q q
       RQV2SetTableCustomFields q -> runSetTableCustomFieldsQV2 q
       RQV2TrackFunction q        -> runTrackFunctionV2 q
-
 
 requiresAdmin :: RQLQuery -> Bool
 requiresAdmin = \case